+++ app/src/main/java/com/samadihadis/countryinformationapplication/data/sub/Currencies.kt
@@ -5,5 +5,5 @@ import java.io.Serializable
 
 data class Currencies(
 
-    @SerializedName("AUD") val AUD: AUD
+    @SerializedName("AUD") val AUD : AUD
 ): Serializable
