+++ app/src/main/java/com/samadihadis/countryinformationapplication/data/sub/NativeName.kt
@@ -1,7 +1,7 @@
 package com.samadihadis.countryinformationapplication.data.sub
 
 import com.google.gson.annotations.SerializedName
-import com.samadihadis.countryinformationapplication.data.language.Eng
+import com.samadihadis.countryinformationapplication.data.translations.Eng
 import java.io.Serializable
 
 data class NativeName(
