+++ app/src/main/java/com/samadihadis/countryinformationapplication/data/sub/Translations.kt
@@ -1,31 +1,31 @@
 package com.samadihadis.countryinformationapplication.data.sub
 
 import com.google.gson.annotations.SerializedName
-import com.samadihadis.countryinformationapplication.data.language.Ara
-import com.samadihadis.countryinformationapplication.data.language.Bre
-import com.samadihadis.countryinformationapplication.data.language.Ces
-import com.samadihadis.countryinformationapplication.data.language.Cym
-import com.samadihadis.countryinformationapplication.data.language.Deu
-import com.samadihadis.countryinformationapplication.data.language.Est
-import com.samadihadis.countryinformationapplication.data.language.Fin
-import com.samadihadis.countryinformationapplication.data.language.Fra
-import com.samadihadis.countryinformationapplication.data.language.Hrv
-import com.samadihadis.countryinformationapplication.data.language.Hun
-import com.samadihadis.countryinformationapplication.data.language.Ita
-import com.samadihadis.countryinformationapplication.data.language.Jpn
-import com.samadihadis.countryinformationapplication.data.language.Kor
-import com.samadihadis.countryinformationapplication.data.language.Nld
-import com.samadihadis.countryinformationapplication.data.language.Per
-import com.samadihadis.countryinformationapplication.data.language.Pol
-import com.samadihadis.countryinformationapplication.data.language.Por
-import com.samadihadis.countryinformationapplication.data.language.Rus
-import com.samadihadis.countryinformationapplication.data.language.Slk
-import com.samadihadis.countryinformationapplication.data.language.Spa
-import com.samadihadis.countryinformationapplication.data.language.Srp
-import com.samadihadis.countryinformationapplication.data.language.Swe
-import com.samadihadis.countryinformationapplication.data.language.Tur
-import com.samadihadis.countryinformationapplication.data.language.Urd
-import com.samadihadis.countryinformationapplication.data.language.Zho
+import com.samadihadis.countryinformationapplication.data.translations.Ara
+import com.samadihadis.countryinformationapplication.data.translations.Bre
+import com.samadihadis.countryinformationapplication.data.translations.Ces
+import com.samadihadis.countryinformationapplication.data.translations.Cym
+import com.samadihadis.countryinformationapplication.data.translations.Deu
+import com.samadihadis.countryinformationapplication.data.translations.Est
+import com.samadihadis.countryinformationapplication.data.translations.Fin
+import com.samadihadis.countryinformationapplication.data.translations.Fra
+import com.samadihadis.countryinformationapplication.data.translations.Hrv
+import com.samadihadis.countryinformationapplication.data.translations.Hun
+import com.samadihadis.countryinformationapplication.data.translations.Ita
+import com.samadihadis.countryinformationapplication.data.translations.Jpn
+import com.samadihadis.countryinformationapplication.data.translations.Kor
+import com.samadihadis.countryinformationapplication.data.translations.Nld
+import com.samadihadis.countryinformationapplication.data.translations.Per
+import com.samadihadis.countryinformationapplication.data.translations.Pol
+import com.samadihadis.countryinformationapplication.data.translations.Por
+import com.samadihadis.countryinformationapplication.data.translations.Rus
+import com.samadihadis.countryinformationapplication.data.translations.Slk
+import com.samadihadis.countryinformationapplication.data.translations.Spa
+import com.samadihadis.countryinformationapplication.data.translations.Srp
+import com.samadihadis.countryinformationapplication.data.translations.Swe
+import com.samadihadis.countryinformationapplication.data.translations.Tur
+import com.samadihadis.countryinformationapplication.data.translations.Urd
+import com.samadihadis.countryinformationapplication.data.translations.Zho
 import java.io.Serializable
 
 data class Translations(
