+++ app/src/main/java/com/samadihadis/countryinformationapplication/data/translations/Ara.kt
@@ -1,9 +1,10 @@
-package com.samadihadis.countryinformationapplication.data.language
+package com.samadihadis.countryinformationapplication.data.translations
 
 import com.google.gson.annotations.SerializedName
 import java.io.Serializable
 
-data class Urd(
+data class Ara(
+
     @SerializedName("official") val official: String,
     @SerializedName("common") val common: String
-): Serializable
+) : Serializable
