+++ app/src/main/java/com/samadihadis/countryinformationapplication/data/translations/Fin.kt
@@ -1,10 +1,10 @@
-package com.samadihadis.countryinformationapplication.data.language
+package com.samadihadis.countryinformationapplication.data.translations
 
 import com.google.gson.annotations.SerializedName
 import java.io.Serializable
 
-data class Ara(
+data class Fin(
 
     @SerializedName("official") val official: String,
     @SerializedName("common") val common: String
-) : Serializable
+): Serializable
