+++ app/src/main/java/com/samadihadis/countryinformationapplication/data/translations/Hrv.kt
@@ -1,9 +1,9 @@
-package com.samadihadis.countryinformationapplication.data.language
+package com.samadihadis.countryinformationapplication.data.translations
 
 import com.google.gson.annotations.SerializedName
 import java.io.Serializable
 
-data class Ces(
+data class Hrv(
 
     @SerializedName("official") val official: String,
     @SerializedName("common") val common: String
