+++ app/src/main/java/com/samadihadis/countryinformationapplication/data/translations/Per.kt
@@ -1,9 +1,9 @@
-package com.samadihadis.countryinformationapplication.data.language
+package com.samadihadis.countryinformationapplication.data.translations
 
 import com.google.gson.annotations.SerializedName
 import java.io.Serializable
 
-data class Bre(
+data class Per(
 
     @SerializedName("official") val official: String,
     @SerializedName("common") val common: String
