+++ app/src/main/java/com/samadihadis/countryinformationapplication/data/translations/Pol.kt
@@ -1,9 +1,9 @@
-package com.samadihadis.countryinformationapplication.data.language
+package com.samadihadis.countryinformationapplication.data.translations
 
 import com.google.gson.annotations.SerializedName
 import java.io.Serializable
 
-data class Nld(
+data class Pol(
 
     @SerializedName("official") val official: String,
     @SerializedName("common") val common: String
