+++ app/src/main/java/com/samadihadis/countryinformationapplication/data/translations/Srp.kt
@@ -1,11 +1,10 @@
-package com.samadihadis.countryinformationapplication.data.language
+package com.samadihadis.countryinformationapplication.data.translations
 
 import com.google.gson.annotations.SerializedName
 import java.io.Serializable
 
-data class Swe(
+data class Srp(
 
     @SerializedName("official") val official: String,
     @SerializedName("common") val common: String
-
 ): Serializable
