+++ app/src/main/java/com/samadihadis/countryinformationapplication/data/translations/Swe.kt
@@ -1,10 +1,11 @@
-package com.samadihadis.countryinformationapplication.data.language
+package com.samadihadis.countryinformationapplication.data.translations
 
 import com.google.gson.annotations.SerializedName
 import java.io.Serializable
 
-data class Tur(
+data class Swe(
 
     @SerializedName("official") val official: String,
     @SerializedName("common") val common: String
+
 ): Serializable
