+++ app/src/main/java/com/samadihadis/countryinformationapplication/data/translations/Urd.kt
@@ -1,10 +1,9 @@
-package com.samadihadis.countryinformationapplication.data.language
+package com.samadihadis.countryinformationapplication.data.translations
 
 import com.google.gson.annotations.SerializedName
 import java.io.Serializable
 
-data class Fra(
-
+data class Urd(
     @SerializedName("official") val official: String,
     @SerializedName("common") val common: String
 ): Serializable
